requirements.txt
aiogram
python-dotenv
playwright
